+++ frontend/src/hooks/useSocket.ts
@@ -10,7 +10,7 @@ const useSocket = () => {
         };
 
         const handleDisconnect = (reason: any) => {
-            console.log("disconnect reason", reason);
+            // console.log("disconnect reason", reason);
             setIsConnected(false);
         };
 
@@ -21,7 +21,7 @@ const useSocket = () => {
         // });
         // Ensures socket reconnects on page refresh
         if (!socket.connected) {
-            console.log("Connecting socket...");
+            // console.log("Connecting socket...");
             socket.connect();
         }